flags
I notice that when I pause a prayer and it is still time for the prayer, it automatically plays again. This is because when i paused the prayer, it was still time for the prayer, the condition to play a prayer is if(it is time for prayer and mp3 player is not busy), so when i pause the prayer the mp3 player is no longer busy and if it is still time for the prayer the condition "if(it is time for prayer and mp3 player is not busy)" becomes true and the prayer plays again. So i added a flag to play the prayer only once